IFCLora/requirements.txt
torch>=2.1.0
transformers>=4.40.0
peft>=0.10.0
accelerate>=0.29.0
bitsandbytes>=0.43.0
networkx>=3.1
numpy>=1.24.0
datasets>=2.18.0
scipy>=1.11.0
pyyaml
loguru
tqdm
trl
